:hammer: Flatten in a way that keeps mypy happy
Bit of a tail-wagging-the-dog situation here, but it seems that mypy can't
grok this use of reduce/concat (there's even an issue for this and it seems
to be outstanding); pyright is happy, but mypy isn't.

So... it doesn't make a whole heap of difference really so here I flatten a
couple of sequences in a different way that mypy can handle.