:sparkles: Add a display of a PEP's details to the detail widget
In retrospect, I don't think I like this approach. I thought it was going to
be quick and simple to use Markdown and, while it is, and while the result
looks fine, there are a couple of issues.

The first is that if this is visible it makes navigation through the list
view really stuttery. The second is that there's a lot of useful information
in the widget that the user might want to interact with, that won't work
with the keyboard (because Textual's Markdown widget is keyboard-hostile).

So I think I'm going to roll with this for now while I get on with other
things, but I think I'll rewrite this soon, perhaps before I finally release
this app.