:construction: WIP reworking the PEP details
Using Markdown to show the details was quick and pragmatic, but on release I
want people to be able to navigate everything that can be interacted with
using the keyboard. So here I'm dropping Markdown and moving to some custom
widgets.

These widgets aren't all finished yet, but I'm at the end of a coding
session and I want to save and commit this for next time.

Also, I'm moving this panel to the right of the display. Having it at the
bottom and height auto was a nice idea but it caused horrific lag as I moved
through the list of PEPs. This looks to be a problem in Textual, but I can't
narrow down the exact circumstance.

Moving the widget to the right so it doesn't disturb the option list seems
to do the tick and I think the result looks fine.