:hammer: Simplify the fields for the PEP details
There's two or three different fields that I want the user to be able to
navigate to, to highlight, to select, or even click on. I could have
custom-made some fields to do this (and I was doing so) but then I realised
that OptionList has everything I need to do that: an OptionList with one
value works just as well in that capacity as anything else I might create.