:lipstick: Don't allow the details panel itself to get focus
This felt like the right thing to do for a while, and it sort of comes from
what I was doing with Tinboard's and Braindrop's UIs, but I don't think it
makes sense here. Here there will always be a widget that can receive focus;
so given this always have focus go *into* the panel.