:fire: Remove some unused PEP details panel styling
This is a hangover from when I was toying with it being a pop-up from the
bottom of the display. This didn't work at all as it caused all sorts of
performance issues with Textual (perhaps one day I'll narrow the cause of
that lag down), so I'm sticking with it being a side panel.