:hammer: Swap to hand-parsing PEP dates
Using strptime with locale-protection was a nice way of doing things, but
relying on locale support being everywhere is too much of a headache for too
little gain. While this was working fine on macOS; it was causing all sorts
of trouble in GitHub's actions using GNU/Linux.

Rather than needing a specific environment to get things done, it makes
sense to take the less-ideal route to parsing a date while keeping it
pragmatic.